third_party/WebKit/LayoutTests/fast/encoding/api/basics.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>

description("This tests the basics of the Encoding API.");

shouldBeTrue("'TextEncoder' in window");
shouldBeTrue("'TextDecoder' in window");

shouldBeTrue("'encoding' in new TextEncoder");
shouldBeTrue("'encoding' in new TextDecoder");

shouldBeEqualToString("typeof (new TextEncoder).encoding", "string");
shouldBeEqualToString("typeof (new TextDecoder).encoding", "string");

shouldBeTrue("'encode' in new TextEncoder");
shouldBeTrue("'decode' in new TextDecoder");

shouldBeEqualToString("typeof (new TextEncoder).encode", "function");
shouldBeEqualToString("typeof (new TextDecoder).decode", "function");


shouldBeEqualToString("(new TextEncoder).encoding", "utf-8");
shouldBeEqualToString("(new TextDecoder).encoding", "utf-8");

function toArray(arrayLike) {
    return [].map.call(arrayLike, function(x) { return x; });
}

function testEncodeDecodeSample(encoding, string, bytes) {
    debug("");
    debug("test encode/decode sample - " + encoding);

    evalAndLog("encoded = new TextEncoder('" + encoding + "').encode(" + JSON.stringify(string) + ")");
    shouldBeEqualToString("JSON.stringify(toArray(encoded))", JSON.stringify(bytes));
    shouldBeEqualToString("new TextDecoder('" + encoding + "').decode(new Uint8Array(" + JSON.stringify(bytes) + "))", string);
}

testEncodeDecodeSample(
    "utf-8",
    "z\xA2\u6C34\uD834\uDD1E\uDBFF\uDFFD", // z, cent, CJK water, G-Clef, Private-use character
    [0x7A, 0xC2, 0xA2, 0xE6, 0xB0, 0xB4, 0xF0, 0x9D, 0x84, 0x9E, 0xF4, 0x8F, 0xBF, 0xBD]
);
testEncodeDecodeSample(
    "utf-16le",
    "z\xA2\u6C34\uD834\uDD1E\uDBFF\uDFFD", // z, cent, CJK water, G-Clef, Private-use character
    [0x7A, 0x00, 0xA2, 0x00, 0x34, 0x6C, 0x34, 0xD8, 0x1E, 0xDD, 0xFF, 0xDB, 0xFD, 0xDF]
);
testEncodeDecodeSample(
    "utf-16be",
    "z\xA2\u6C34\uD834\uDD1E\uDBFF\uDFFD", // z, cent, CJK water, G-Clef, Private-use character
    [0x00, 0x7A, 0x00, 0xA2, 0x6C, 0x34, 0xD8, 0x34, 0xDD, 0x1E, 0xDB, 0xFF, 0xDF, 0xFD]
);
testEncodeDecodeSample(
    "utf-16",
    "z\xA2\u6C34\uD834\uDD1E\uDBFF\uDFFD", // z, cent, CJK water, G-Clef, Private-use character
    [0x7A, 0x00, 0xA2, 0x00, 0x34, 0x6C, 0x34, 0xD8, 0x1E, 0xDD, 0xFF, 0xDB, 0xFD, 0xDF]
);

</script>
